MRI/requirements.txt
nilearn==0.10.1
scikit-learn==1.0.2
matplotlib==3.3.1
seaborn==0.12.2
scipy==1.10.1
pandas==2.0.2
fracridge==1.3.1
joblib==1.2.0
tqdm==4.56.0
pybids==0.8.0
nipype==1.8.6
cython==0.29.22
h5py==3.10.0
pycortex==1.2.7
cmasher==1.6.3
duecredit==0.9.1
jupyter
